attr: isolate terios methods into platform-specific file
